Refactor: 파일명 변경 및 console 주석 처리
