+++ src/app/mj/page.tsx
@@ -3,7 +3,7 @@
 import WriteEditModal from '@/components/WriteEditModal/WriteEditModal';
 import { WriteEditModalType, WRITE_EIDT_MODAL_TYPE } from '@/constants/writeEditModalType';
 import { useState } from 'react';
-import { BUTTON_COLOR } from '@/constants/buttonTypes';
+import { BUTTON_COLOR } from '@/constants/buttonColors';
 import classNames from 'classnames/bind';
 import Button from '@/components/Button/Button';
 import styles from './page.module.scss';
@@ -30,7 +30,7 @@ export default function Page() {
 
   const handleButtonClick = () => {
     /** 버튼 누를 때 실행되는 함수 */
-    console.log('버튼 누름');
+    // console.log('버튼 누름');
   };
   return (
     <div className={cn('container')}>
